refactor : 도메인 model 패키지 구조 변경
