+++ src/main/java/org/ezcode/codetest/application/community/dto/request/DiscussionCreateRequest.java
@@ -2,7 +2,7 @@ package org.ezcode.codetest.application.community.dto.request;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.*;
 
-import org.ezcode.codetest.domain.community.model.Discussion;
+import org.ezcode.codetest.domain.community.model.entity.Discussion;
 import org.ezcode.codetest.domain.language.model.entity.Language;
 import org.ezcode.codetest.domain.problem.model.entity.Problem;
 import org.ezcode.codetest.domain.user.model.entity.User;
