+++ src/main/java/org/ezcode/codetest/application/community/service/DiscussionService.java
@@ -3,7 +3,7 @@ package org.ezcode.codetest.application.community.service;
 import org.ezcode.codetest.application.community.dto.request.DiscussionCreateRequest;
 import org.ezcode.codetest.application.community.dto.request.DiscussionModifyRequest;
 import org.ezcode.codetest.application.community.dto.response.DiscussionResponse;
-import org.ezcode.codetest.domain.community.model.Discussion;
+import org.ezcode.codetest.domain.community.model.entity.Discussion;
 import org.ezcode.codetest.domain.community.service.DiscussionDomainService;
 import org.ezcode.codetest.domain.language.model.entity.Language;
 import org.ezcode.codetest.domain.problem.model.entity.Problem;
