+++ src/main/java/org/ezcode/codetest/application/community/service/DiscussionVoteService.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.ezcode.codetest.application.community.dto.response.VoteResponse;
 import org.ezcode.codetest.application.notification.event.NotificationCreateEvent;
 import org.ezcode.codetest.application.notification.port.NotificationEventService;
-import org.ezcode.codetest.domain.community.model.Discussion;
-import org.ezcode.codetest.domain.community.model.DiscussionVote;
+import org.ezcode.codetest.domain.community.model.entity.Discussion;
+import org.ezcode.codetest.domain.community.model.entity.DiscussionVote;
 import org.ezcode.codetest.domain.community.service.DiscussionDomainService;
 import org.ezcode.codetest.domain.community.service.DiscussionVoteDomainService;
 import org.ezcode.codetest.domain.user.model.entity.User;
