+++ src/main/java/org/ezcode/codetest/application/community/service/ReplyVoteService.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.ezcode.codetest.application.community.dto.response.VoteResponse;
 import org.ezcode.codetest.application.notification.event.NotificationCreateEvent;
 import org.ezcode.codetest.application.notification.port.NotificationEventService;
-import org.ezcode.codetest.domain.community.model.Reply;
-import org.ezcode.codetest.domain.community.model.ReplyVote;
+import org.ezcode.codetest.domain.community.model.entity.Reply;
+import org.ezcode.codetest.domain.community.model.entity.ReplyVote;
 import org.ezcode.codetest.domain.community.service.ReplyDomainService;
 import org.ezcode.codetest.domain.community.service.ReplyVoteDomainService;
 import org.ezcode.codetest.domain.user.model.entity.User;
