+++ src/main/java/org/ezcode/codetest/domain/community/model/entity/Reply.java
@@ -1,4 +1,4 @@
-package org.ezcode.codetest.domain.community.model;
+package org.ezcode.codetest.domain.community.model.entity;
 
 import java.util.ArrayList;
 import java.util.HashSet;
