+++ src/main/java/org/ezcode/codetest/domain/community/repository/DiscussionVoteRepository.java
@@ -1,6 +1,6 @@
 package org.ezcode.codetest.domain.community.repository;
 
-import org.ezcode.codetest.domain.community.model.DiscussionVote;
+import org.ezcode.codetest.domain.community.model.entity.DiscussionVote;
 
 public interface DiscussionVoteRepository extends BaseVoteRepository<DiscussionVote> {
 }
