+++ src/main/java/org/ezcode/codetest/domain/community/repository/ReplyRepository.java
@@ -2,7 +2,7 @@ package org.ezcode.codetest.domain.community.repository;
 
 import java.util.Optional;
 
-import org.ezcode.codetest.domain.community.model.Reply;
+import org.ezcode.codetest.domain.community.model.entity.Reply;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
