+++ src/main/java/org/ezcode/codetest/domain/community/repository/ReplyVoteRepository.java
@@ -1,6 +1,6 @@
 package org.ezcode.codetest.domain.community.repository;
 
-import org.ezcode.codetest.domain.community.model.ReplyVote;
+import org.ezcode.codetest.domain.community.model.entity.ReplyVote;
 
 public interface ReplyVoteRepository extends BaseVoteRepository<ReplyVote> {
 }
