+++ src/main/java/org/ezcode/codetest/domain/community/service/DiscussionDomainService.java
@@ -2,7 +2,7 @@ package org.ezcode.codetest.domain.community.service;
 
 import org.ezcode.codetest.domain.community.exception.CommunityException;
 import org.ezcode.codetest.domain.community.exception.CommunityExceptionCode;
-import org.ezcode.codetest.domain.community.model.Discussion;
+import org.ezcode.codetest.domain.community.model.entity.Discussion;
 import org.ezcode.codetest.domain.community.repository.DiscussionRepository;
 import org.ezcode.codetest.domain.language.model.entity.Language;
 import org.springframework.data.domain.Page;
