+++ src/main/java/org/ezcode/codetest/domain/community/service/DiscussionVoteDomainService.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.ezcode.codetest.application.notification.enums.NotificationType;
 import org.ezcode.codetest.application.notification.event.NotificationCreateEvent;
 import org.ezcode.codetest.application.notification.event.payload.DiscussionVotePayload;
-import org.ezcode.codetest.domain.community.model.Discussion;
-import org.ezcode.codetest.domain.community.model.DiscussionVote;
+import org.ezcode.codetest.domain.community.model.entity.Discussion;
+import org.ezcode.codetest.domain.community.model.entity.DiscussionVote;
 import org.ezcode.codetest.domain.community.repository.DiscussionVoteRepository;
 import org.ezcode.codetest.domain.user.model.entity.User;
 import org.springframework.stereotype.Service;
