+++ src/main/java/org/ezcode/codetest/domain/community/service/ReplyVoteDomainService.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 import org.ezcode.codetest.application.notification.enums.NotificationType;
 import org.ezcode.codetest.application.notification.event.NotificationCreateEvent;
 import org.ezcode.codetest.application.notification.event.payload.ReplyVotePayload;
-import org.ezcode.codetest.domain.community.model.Discussion;
-import org.ezcode.codetest.domain.community.model.Reply;
-import org.ezcode.codetest.domain.community.model.ReplyVote;
+import org.ezcode.codetest.domain.community.model.entity.Discussion;
+import org.ezcode.codetest.domain.community.model.entity.Reply;
+import org.ezcode.codetest.domain.community.model.entity.ReplyVote;
 import org.ezcode.codetest.domain.community.repository.ReplyVoteRepository;
 import org.ezcode.codetest.domain.user.model.entity.User;
 import org.springframework.stereotype.Service;
