+++ src/main/java/org/ezcode/codetest/infrastructure/persistence/repository/community/discussion/DiscussionRepositoryImpl.java
@@ -2,7 +2,7 @@ package org.ezcode.codetest.infrastructure.persistence.repository.community.disc
 
 import java.util.Optional;
 
-import org.ezcode.codetest.domain.community.model.Discussion;
+import org.ezcode.codetest.domain.community.model.entity.Discussion;
 import org.ezcode.codetest.domain.community.repository.DiscussionRepository;
 import org.ezcode.codetest.domain.language.model.entity.Language;
 import org.springframework.data.domain.Page;
