+++ src/main/java/org/ezcode/codetest/infrastructure/persistence/repository/community/vote/DiscussionVoteJpaRepository.java
@@ -2,7 +2,7 @@ package org.ezcode.codetest.infrastructure.persistence.repository.community.vote
 
 import java.util.Optional;
 
-import org.ezcode.codetest.domain.community.model.DiscussionVote;
+import org.ezcode.codetest.domain.community.model.entity.DiscussionVote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiscussionVoteJpaRepository extends JpaRepository<DiscussionVote, Long> {
