--- src/main/java/org/ezcode/codetest/presentation/LikeController.java
@@ -1,4 +1,0 @@
-package org.ezcode.codetest.presentation;
-
-public class LikeController {
-}
